Fix: satellite service upgrade file lock (#352)
This PR aims to resolve issue related to Satellite service still running during the upgrade, which results in the files being used/locked and not being updated.
I haven't found a way to replicate this reliably but user reports have confirmed that this might be the case.

In addition, this PR includes more file removal attempts for the Libre Hardware Monitor dll as well as the "HASS.Agent.Satellite.Service.sys" file that is being picked up by Windows Defender.