Fearure: monitor sleep based on power plan modification (#380)
This PR adds feature requested via #305 - command that for a short period of time modifies the power plan to make the screens turn off.
The power plan is restored after command finishes execution.

This is an attempt to provide a way to turn of monitors on systems that only have the "modern/connected sleep" provided by the manufacturers.